% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbasis.R
\name{bbasis}
\alias{bbasis}
\title{Construct B-spline basis}
\usage{
bbasis(x, xl, xr, ndx, bdeg)
}
\arguments{
\item{x}{A numerical vector.}

\item{xl}{A numerical value that indicates a left boundary for \code{x}.}

\item{xr}{A numerical value that indicates a right boundary for \code{x}.}

\item{ndx}{A numerical value that indicates the number of internal intervals (or the number of internal knots minus one).}

\item{bdeg}{A numerical value that indicates the degree of the B-splines.}
}
\value{
A list that contains the B-spline basis (\code{B}), and the internal knots (\code{knots}).
}
\description{
It constructs a B-spline basis of equally-spaced knots from a (numerical) covariate within the function \code{mm_basis}.
}
\examples{
x <- seq(0, 1, length = 200)
xls <- min(x) - 0.001
xrs <- max(x) + 0.001
ndxs <- 10
## A quadratic B-spline basis
B2 <- bbasis(x, xl = xls, xr = xrs, ndx = ndxs, bdeg = 2)$B
matplot(x, B2, t = "l", lty = 1, lwd = 2, col = rainbow(ncol(B2)))
## A cubic B-spline basis
B3 <- bbasis(x, xl = xls, xr = xrs, ndx = ndxs, bdeg = 3)$B
matplot(x, B3, t = "l", lty = 1, lwd = 2, col = rainbow(ncol(B3)))
}
\seealso{
\code{\link{tpower}}, \code{\link{mm_basis}}
}
