% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bblock2.R
\name{bblock2}
\alias{bblock2}
\title{Construct matrix from two-by-two blocks}
\usage{
bblock2(A11, A12, A21, A22)
}
\arguments{
\item{A11, A12, A21, A22}{blocks of the partitioned matrix}
}
\value{
The assembled matrix \code{rbind(cbind(A11, A12), cbind(A21, A22))}.
}
\description{
This function recovers a partitioned matrix from two-by-two blocks.
}
\keyword{internal}
