% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{btWb_cpp}
\alias{btWb_cpp}
\title{Form B' diag(w) B + apply (dense) for Camarda update RHS B'(w*z)}
\usage{
btWb_cpp(B, w, z)
}
\description{
Form B' diag(w) B + apply (dense) for Camarda update RHS B'(w*z)
}
\keyword{internal}
