% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{clmm_crossprod_cpp}
\alias{clmm_crossprod_cpp}
\title{Working cross-products for CLMM given CGX, CGZ (coarse x p / q)}
\usage{
clmm_crossprod_cpp(CGX, CGZ, z, w)
}
\description{
Working cross-products for CLMM given CGX, CGZ (coarse x p / q)
}
\keyword{internal}
