% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{comp_mul_groups_cpp}
\alias{comp_mul_groups_cpp}
\title{Partition (rowsum) aggregation: C \%*\% (gamma * A) for 0-1 C}
\usage{
comp_mul_groups_cpp(gamma, A, groups, n_coarse)
}
\arguments{
\item{gamma}{length n_fine}

\item{A}{n_fine x p}

\item{groups}{1-based group id length n_fine}

\item{n_coarse}{number of coarse cells}
}
\description{
Compiled helper for 0-1 composition matrices.
}
\keyword{internal}
