% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.as_clgam}
\alias{.as_clgam}
\title{Attach S3 class; \code{family} is the GLM family, \code{type} is spatial/contrast.}
\usage{
.as_clgam(x, call = NULL, type = c("spatial", "contrast"), family = "poisson")
}
\description{
Attach S3 class; \code{family} is the GLM family, \code{type} is spatial/contrast.
}
\keyword{internal}
