% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_backend.R
\name{.as_spam}
\alias{.as_spam}
\title{Coerce Matrix / matrix to spam without relying on missing S4 methods}
\usage{
.as_spam(C)
}
\description{
Coerce Matrix / matrix to spam without relying on missing S4 methods
}
\keyword{internal}
