% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.build_orth_Af}
\alias{.build_orth_Af}
\title{Design space used to orthogonalize spatial bases: fine B-splines + linear fine.
Coarse-scale smooths (Case B/C aggregated terms) are excluded by design.}
\usage{
.build_orth_Af(Bk = NULL, nl.level = NULL, lcovfine = NULL)
}
\description{
Design space used to orthogonalize spatial bases: fine B-splines + linear fine.
Coarse-scale smooths (Case B/C aggregated terms) are excluded by design.
}
\keyword{internal}
