% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clgam-methods.R
\name{.clgam_col_ramp}
\alias{.clgam_col_ramp}
\title{Map numeric values to colours}
\usage{
.clgam_col_ramp(v, palette = "Plasma", n = 100L)
}
\description{
Map numeric values to colours
}
\keyword{internal}
