% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clgam-formula.R
\name{.clgam_collect_s}
\alias{.clgam_collect_s}
\title{Walk \code{+} / \code{s()} on a formula RHS}
\usage{
.clgam_collect_s(expr, env)
}
\description{
Walk \code{+} / \code{s()} on a formula RHS
}
\keyword{internal}
