% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clgam-methods.R
\name{.clgam_colorbar}
\alias{.clgam_colorbar}
\title{Draw a vertical colour bar in the right margin}
\usage{
.clgam_colorbar(zlim, palette, title = NULL)
}
\description{
Draw a vertical colour bar in the right margin
}
\keyword{internal}
