% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clgam-contrast-infer.R
\name{.clgam_contrast_uncond_se}
\alias{.clgam_contrast_uncond_se}
\title{Wood–Pya–Säfken inflate of a contrast linear functional.}
\usage{
.clgam_contrast_uncond_se(fit, type, eps = 0.001)
}
\description{
Wood–Pya–Säfken inflate of a contrast linear functional.
}
\keyword{internal}
