% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-clgam.R
\name{.clgam_edf_names}
\alias{.clgam_edf_names}
\title{SOP term names: spatial penalties then smooth covariates.}
\usage{
.clgam_edf_names(fit)
}
\description{
SOP term names: spatial penalties then smooth covariates.
}
\keyword{internal}
