% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-clgam.R
\name{.clgam_infer_converged}
\alias{.clgam_infer_converged}
\title{Infer convergence for current and older fit objects.}
\usage{
.clgam_infer_converged(object)
}
\description{
Infer convergence for current and older fit objects.
}
\keyword{internal}
