% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.clgam_mat_colnames}
\alias{.clgam_mat_colnames}
\title{Column names for a covariate matrix; fill missing with \code{prefix} + index.}
\usage{
.clgam_mat_colnames(M, prefix = "s")
}
\description{
Column names for a covariate matrix; fill missing with \code{prefix} + index.
}
\keyword{internal}
