% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_TMB.R
\name{.clgam_mm_design}
\alias{.clgam_mm_design}
\title{Build mixed-model design (shared by SOP / TMB)}
\usage{
.clgam_mm_design(
  x1,
  x2,
  efine = NULL,
  lcovfine = NULL,
  nlcovfine = NULL,
  C,
  x1lim = NULL,
  x2lim = NULL,
  ndx = c(15, 15),
  bdeg = c(3, 3),
  pord = c(2, 2),
  decom = 1,
  ndxnl = 15,
  bdegnl = 3,
  pordnl = 2,
  nl.basis = "pspline",
  orth.smooth = TRUE,
  sparse.backend = "auto"
)
}
\description{
Build mixed-model design (shared by SOP / TMB)
}
\keyword{internal}
