% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clgam-methods.R
\name{.clgam_plot_contrast_class}
\alias{.clgam_plot_contrast_class}
\title{Three-class contrast map using joint SE of eta1 - eta2}
\usage{
.clgam_plot_contrast_class(x, dots, sf_fine = NULL, level = 0.95)
}
\description{
Three-class contrast map using joint SE of eta1 - eta2
}
\keyword{internal}
