% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re.R
\name{.clgam_resolve_re}
\alias{.clgam_resolve_re}
\title{Resolve the optional coarse-level random effect}
\usage{
.clgam_resolve_re(re = NULL)
}
\arguments{
\item{re}{\code{NULL}/\code{"none"} (default) or \code{"coarse"}}
}
\value{
\code{"none"} or \code{"coarse"}
}
\description{
Resolve the optional coarse-level random effect
}
\keyword{internal}
