% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_backend.R
\name{.clgam_sparse_backend}
\alias{.clgam_sparse_backend}
\title{Default sparse backend: \code{"Matrix"}, \code{"spam"}, or \code{"dense"}}
\usage{
.clgam_sparse_backend(backend = NULL)
}
\description{
Default sparse backend: \code{"Matrix"}, \code{"spam"}, or \code{"dense"}
}
\keyword{internal}
