% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-clgam.R
\name{.clgam_term_table}
\alias{.clgam_term_table}
\title{Term table: tau^2 / SOP ED / kappa by covariate (linear rows have no tau^2).}
\usage{
.clgam_term_table(object, kappa_by = NULL)
}
\description{
Term table: tau^2 / SOP ED / kappa by covariate (linear rows have no tau^2).
}
\keyword{internal}
