% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_ata.R
\name{.clgam_unit_square}
\alias{.clgam_unit_square}
\title{Unit-square envelope as an \pkg{sf} polygon}
\usage{
.clgam_unit_square()
}
\description{
Unit-square envelope as an \pkg{sf} polygon
}
\keyword{internal}
