% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re.R
\name{.clgam_Z_re}
\alias{.clgam_Z_re}
\title{Incidence matrix mapping fine units to coarse cells (partition C)}
\usage{
.clgam_Z_re(groups, n_coarse)
}
\description{
Incidence matrix mapping fine units to coarse cells (partition C)
}
\keyword{internal}
