% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clmm_mat.R
\name{.clmm_working_z}
\alias{.clmm_working_z}
\title{Working response for composite-link Poisson PIRLS}
\usage{
.clmm_working_z(C, gamma, eta, mu, y, groups = NULL)
}
\description{
Working response for composite-link Poisson PIRLS
}
\keyword{internal}
