% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_backend.R
\name{.comp_mul}
\alias{.comp_mul}
\title{C %*% (gamma * A) with Matrix, spam, or partition rowsum}
\usage{
.comp_mul(C, gamma, A, groups = NULL)
}
\description{
C %*% (gamma * A) with Matrix, spam, or partition rowsum
}
\keyword{internal}
