% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.orth_cols}
\alias{.orth_cols}
\title{Column-wise residualize A against span(B) (QR). Drop near-zero columns of B.}
\usage{
.orth_cols(A, B)
}
\description{
Column-wise residualize A against span(B) (QR). Drop near-zero columns of B.
}
\keyword{internal}
