% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_backend.R
\name{.partition_groups}
\alias{.partition_groups}
\title{Fine->coarse group id for partition C (length = ncol(C))}
\usage{
.partition_groups(C)
}
\description{
Fine->coarse group id for partition C (length = ncol(C))
}
\keyword{internal}
