% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.pearson_phi}
\alias{.pearson_phi}
\title{Pearson dispersion for (quasi-)Poisson composite-link fits.}
\usage{
.pearson_phi(y, mu, ed)
}
\description{
\deqn{\hat\phi = \frac{1}{n-p}\sum_i (y_i-\hat\mu_i)^2 / \hat\mu_i}
with \eqn{p} the effective dimension (\code{ed} when available).
}
\keyword{internal}
