% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.scale_se_phi}
\alias{.scale_se_phi}
\title{Inflate stored SE fields by \eqn{\sqrt{\phi}} (quasi-Poisson).}
\usage{
.scale_se_phi(out, phi)
}
\description{
Inflate stored SE fields by \eqn{\sqrt{\phi}} (quasi-Poisson).
}
\keyword{internal}
