% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sop_kron_spatial.R
\name{.sop_ginv_b3}
\alias{.sop_ginv_b3}
\title{B3 slice of the spatial G-inverse diagonal}
\usage{
.sop_ginv_b3(meta, la)
}
\description{
B3 slice of the spatial G-inverse diagonal
}
\keyword{internal}
