% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sop_kron_spatial.R
\name{.sop_ginv_spatial}
\alias{.sop_ginv_spatial}
\title{Spatial G-inverse diagonal matching \code{pois_SOP} inner SOP}
\usage{
.sop_ginv_spatial(meta, la)
}
\arguments{
\item{meta}{output of \code{\link{.sop_kron_meta}}}

\item{la}{length-2 spatial variance components \code{c(tau1^2, tau2^2)}}
}
\value{
numeric vector of length \code{meta$q_spatial}
}
\description{
\code{Ginv = c((1/la[2])*g2u, (1/la[1])*g1u, (1/la[2])*g2b + (1/la[1])*g1b)}.
Nonlinear blocks are not included; append them as in \code{pois_SOP}.
}
\keyword{internal}
