% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sop_kron_spatial.R
\name{.sop_kron_applicable}
\alias{.sop_kron_applicable}
\title{Whether B3 block inverse applies to this \code{meta} / \code{q}}
\usage{
.sop_kron_applicable(meta, q)
}
\description{
Whether B3 block inverse applies to this \code{meta} / \code{q}
}
\keyword{internal}
