% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sop_kron_spatial.R
\name{.sop_pcg_jacobi}
\alias{.sop_pcg_jacobi}
\title{Jacobi-preconditioned CG for \eqn{(N + \mathrm{diag}(g))\theta = \mathrm{rhs}}}
\usage{
.sop_pcg_jacobi(
  N,
  ginv,
  rhs,
  Mdiag = NULL,
  tol = 1e-10,
  maxit = NULL,
  x0 = NULL
)
}
\description{
Jacobi-preconditioned CG for \eqn{(N + \mathrm{diag}(g))\theta = \mathrm{rhs}}
}
\keyword{internal}
