% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sop_kron_spatial.R
\name{.sop_Sinv_dense}
\alias{.sop_Sinv_dense}
\title{Dense inverse of the Schur complement \code{S}}
\usage{
.sop_Sinv_dense(N, G)
}
\description{
Dense inverse of the Schur complement \code{S}
}
\keyword{internal}
