% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv_bblock2.R
\name{inv_bblock2}
\alias{inv_bblock2}
\title{Inverse of a two-by-two nonsingular partitioned matrix}
\usage{
inv_bblock2(A11, A12, A21, A22)
}
\arguments{
\item{A11, A12, A21, A22}{blocks of the partitioned matrix}
}
\value{
list with \code{S11}, \code{S12}, \code{S21}, \code{S22}, \code{S}
}
\description{
Inverse of a two-by-two nonsingular partitioned matrix
}
\keyword{internal}
