% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logdetsym.R
\name{logdetsym}
\alias{logdetsym}
\title{Log-determinant of a symmetric matrix}
\usage{
logdetsym(m)
}
\arguments{
\item{m}{a symmetric numeric matrix}
}
\value{
a scalar, the log of the determinant
}
\description{
Numerically preferable to \code{log(det(m))}.
}
\keyword{internal}
