% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_basis.R
\name{mm_basis}
\alias{mm_basis}
\title{Construct elements of the mixed model representation}
\usage{
mm_basis(x, xl, xr, ndx, bdeg, pord, decom = 1)
}
\arguments{
\item{x}{A numerical vector.}

\item{xl}{A numerical value that indicates a left boundary for \code{x}.}

\item{xr}{A numerical value that indicates a right boundary for \code{x}.}

\item{ndx}{A numerical value that indicates the number of internal intervals (or the number of internal knots minus one).}

\item{bdeg}{A numerical value that indicates the degree of the B-splines.}

\item{pord}{penalty order (number of differences in \code{D})}

\item{decom}{fixed-effect construction: \code{1} uses the penalty
null space (\code{B \%*\% Un}); \code{2} uses the polynomial
\code{1, x, ..., x^(pord-1)}}
}
\value{
A list with fixed \code{X}, random \code{Z}, penalty
  eigenvalues \code{d}, basis \code{B}, dimension \code{m},
  difference matrix \code{D}, eigenvectors \code{U}, and \code{knots}.
}
\description{
Construct elements of the mixed model representation
}
\seealso{
\code{\link{bbasis}}
}
