% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clgam-methods.R
\name{plot.clgam}
\alias{plot.clgam}
\title{Plot a CL-GAMM fit}
\usage{
\method{plot}{clgam}(
  x,
  which = 1:2,
  ask = prod(graphics::par("mfcol")) < length(which) && grDevices::dev.interactive(),
  sf_fine = NULL,
  sf_coarse = NULL,
  g_true = NULL,
  ...
)
}
\arguments{
\item{x}{a \code{clgam} fit}

\item{which}{which plot(s): \code{1} fine \eqn{\hat\eta}; \code{2} y vs
\eqn{\hat\mu}; \code{3} residuals; \code{4} smooth; \code{5} contrast;
\code{6} coarse map of \eqn{y-\hat\mu}; \code{7} classified contrast
(\eqn{\widehat{\eta}_1-\widehat{\eta}_2} vs \code{sd.dif})}

\item{ask}{ask before each plot when length(which) > 1}

\item{sf_fine, sf_coarse}{optional \pkg{sf} layers (else \code{x$sf_fine} /
\code{x$sf_coarse})}

\item{g_true}{optional true smooth vector aligned with fine cells}

\item{...}{graphical parameters (\code{cex}, \code{pch}, \code{lwd}, ...)}
}
\description{
Plot a CL-GAMM fit
}
