% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rten2.R
\name{rten2}
\alias{rten2}
\title{Row-wise or Box Kronecker product of two matrices}
\usage{
rten2(X1, X2)
}
\arguments{
\item{X1}{TODO}

\item{X2}{TODO}
}
\value{
TODO
}
\description{
Row-wise or Box Kronecker product of two matrices
}
