% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_backend.R
\name{sparse_P2D}
\alias{sparse_P2D}
\title{Anisotropic 2D P-spline precision (Kronecker), SparseMatrix}
\usage{
sparse_P2D(
  m1,
  m2,
  lambda1 = 1,
  lambda2 = 1,
  pord = 2L,
  backend = c("Matrix", "spam")
)
}
\arguments{
\item{m1, m2}{basis sizes along each axis}

\item{lambda1, lambda2}{precision weights (e.g. \code{1/tau^2})}

\item{pord}{difference order}

\item{backend}{\code{"Matrix"} (default) or \code{"spam"}}
}
\value{
sparse precision of size \code{(m1*m2) x (m1*m2)}
}
\description{
\deqn{P = \lambda_1 (I_{m_2} \otimes D_1'D_1) + \lambda_2 (D_2'D_2 \otimes I_{m_1})}
}
