% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bblock2.R
\name{bblock2}
\alias{bblock2}
\title{Construct matrix from two-by-two blocks}
\usage{
bblock2(A11, A12, A21, A22)
}
\arguments{
\item{A11}{TODO}

\item{A12}{TODO}

\item{A21}{TODO}

\item{A22}{TODO}
}
\value{
TODO
}
\description{
This function recovers a partitioned matrix from two-by-two blocks.
}
\keyword{internal}
