% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bblock3.R
\name{bblock3}
\alias{bblock3}
\title{Construct matrix from three-by-three blocks}
\usage{
bblock3(A11, A12, A13, A21, A22, A23, A31, A32, A33)
}
\arguments{
\item{A11, A12, A13, A21, A22, A23, A31, A32, A33}{blocks of the partitioned matrix}
}
\value{
The assembled three-by-three block matrix.
}
\description{
This function recovers a partitioned matrix from three-by-three blocks.
}
\keyword{internal}
