% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clgam-methods.R
\name{clgam-methods}
\alias{clgam-methods}
\alias{coef.clgam}
\alias{fitted.clgam}
\alias{residuals.clgam}
\alias{nobs.clgam}
\alias{deviance.clgam}
\alias{logLik.clgam}
\alias{AIC.clgam}
\alias{BIC.clgam}
\title{Extractor and plotting methods for \code{clgam} fits}
\usage{
\method{coef}{clgam}(object, which = c("all", "fixed", "random"), ...)

\method{fitted}{clgam}(object, type = c("mu", "eta", "gamma", "response", "exp.eta"), ...)

\method{residuals}{clgam}(object, type = c("deviance", "pearson", "response"), ...)

\method{nobs}{clgam}(object, ...)

\method{deviance}{clgam}(object, ...)

\method{logLik}{clgam}(object, ...)

\method{AIC}{clgam}(object, ..., k = 2)

\method{BIC}{clgam}(object, ...)
}
\arguments{
\item{object}{a \code{clgam} fit}

\item{...}{unused}

\item{type}{residual type: deviance, Pearson, or response (\code{y - mu})}

\item{k}{penalty in AIC (default 2)}
}
\description{
Extractor and plotting methods for \code{clgam} fits
}
