% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clgam-package.R
\docType{package}
\name{clgam-package}
\alias{clgam-package}
\title{Composite Link Generalized Additive (Mixed) Models}
\description{
Fits CL-GAM / CL-GAMM models for nested areal count disaggregation with
P-splines and separation of overlapping precision matrices (SOP).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/idaejin/CL-GAM}
  \item Report bugs at \url{https://github.com/idaejin/CL-GAM/issues}
}

}
\author{
\strong{Maintainer}: Dae-Jin Lee \email{dae-jin.lee@ie.edu}

Authors:
\itemize{
  \item Dae-Jin Lee \email{dae-jin.lee@ie.edu}
  \item Diego Ayma Anza \email{dayma@est-econ.uc3m.es}
}

Other contributors:
\itemize{
  \item Maria Durban [contributor]
  \item Paul H. C. Eilers [contributor]
}

}
\keyword{internal}
