% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.as_clgam}
\alias{.as_clgam}
\title{Attach S3 class and optional quiet fit banner}
\usage{
.as_clgam(x, call = NULL, family = c("spatial", "contrast"))
}
\description{
Attach S3 class and optional quiet fit banner
}
\keyword{internal}
