% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.ginvsp_from_la}
\alias{.ginvsp_from_la}
\title{Diagonal penalty inverse vector ginvsp for spatial + optional nl smooths.}
\usage{
.ginvsp_from_la(la, g2u, g1u, g2b, g1b, dk = NULL)
}
\description{
Diagonal penalty inverse vector ginvsp for spatial + optional nl smooths.
}
\keyword{internal}
