% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_basis.R
\name{mm_basis}
\alias{mm_basis}
\title{Construct elements of the mixed model representation}
\usage{
mm_basis(x, xl, xr, ndx, bdeg, pord, decom = 1)
}
\arguments{
\item{x}{A numerical vector.}

\item{xl}{A numerical value that indicates a left boundary for \code{x}.}

\item{xr}{A numerical value that indicates a right boundary for \code{x}.}

\item{ndx}{A numerical value that indicates the number of internal intervals (or the number of internal knots minus one).}

\item{bdeg}{A numerical value that indicates the degree of the B-splines.}

\item{pord}{TODO}

\item{decom}{TODO}
}
\value{
TODO
}
\description{
Construct elements of the mixed model representation
}
\seealso{
\code{\link{bbasis}}
}
