% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sop_solve_schur_cpp}
\alias{sop_solve_schur_cpp}
\title{Schur SOP solve + ED diagonal (Ayma nonsymmetric system)}
\usage{
sop_solve_schur_cpp(XtX, ZtX, ZtZ, ZtXtZ, u, G, A11inv_cached)
}
\arguments{
\item{XtX}{p x p}

\item{ZtX}{q x p}

\item{ZtZ}{q x q}

\item{ZtXtZ}{(p+q) x q}

\item{u}{length p+q}

\item{G}{length q (1/Ginv)}

\item{A11inv_cached}{optional p x p inverse of XtX (empty = compute)}
}
\value{
list b_fixed, b_random, dZtNZ, A11inv
}
\description{
Schur SOP solve + ED diagonal (Ayma nonsymmetric system)
}
\keyword{internal}
