% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trprod.R
\name{trprod}
\alias{trprod}
\title{Trace of the product of two matrices}
\usage{
trprod(A, B)
}
\arguments{
\item{A}{TODO}

\item{B}{TODO}
}
\value{
TODO
}
\description{
This function efficiently computes the trace of the product of two matrices \code{A} and \code{B}.
}
\keyword{internal}
